requirements.txt
absl-py==1.0.0
astunparse==1.6.3
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
dill==0.3.4
django-environ==0.9.0
Flask==2.1.2
flatbuffers==2.0
fonttools==4.29.1
fore==0.1
gast==0.5.3
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
halo==0.0.31
html5lib==1.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.8.0rc1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
log-symbols==0.0.14
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
mpmath==1.2.1
ntlm-auth==1.5.0
numpy==1.22.1
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pbr==5.8.1
Pillow==9.0.0
ply==3.11
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pylatexenc==2.10
pyparsing==3.0.7
python-constraint==1.4.0
python-dateutil==2.8.2
python-wordpress-xmlrpc==2.3
qiskit==0.34.2
qiskit-aer==0.10.3
qiskit-ibmq-provider==0.18.3
qiskit-ignis==0.7.0
qiskit-terra==0.19.2
requests==2.27.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
retworkx==0.11.0
rsa==4.8
scipy==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
spinners==0.0.24
stdiomask==0.0.6
stevedore==3.5.0
sympy==1.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0rc1
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tweedledum==1.1.1
typing_extensions==4.0.1
urllib3==1.26.8
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
wrapt==1.13.3
